move parameter test cases to new file
