+++ modules/lily-compiler/src/test/java/io/github/tomboyo/lily/compiler/feature/FluentApiTest.java
@@ -66,110 +66,4 @@ public class FluentApiTest {
           is(URI.create("https://example.com/pets/some-uuid-here")));
     }
   }
-
-  /** Lily generates fluent-style setters for all path parameters. */
-  @Nested
-  class HasFluentPathParameters {
-    private static String packageName;
-
-    @BeforeAll
-    static void beforeAll() throws Exception {
-      packageName =
-          compileOas(
-              """
-              openapi: 3.0.2
-              paths:
-                /pets/{id}:
-                  get:
-                    operationId: getPetById
-                    parameters:
-                      - name: id
-                        in: path
-                        required: true
-                        schema:
-                          type: string
-              """);
-    }
-
-    @Test
-    void hasPathParameterSetters() {
-      var actual =
-          evaluate(
-              """
-              return %s.Api.newBuilder()
-                .uri("https://example.com/")
-                .build()
-                .everyOperation()
-                .getPetById()
-                .setId("1234")
-                .uriTemplate()
-                .toURI();
-              """
-                  .formatted(packageName),
-              URI.class);
-      assertThat(
-          "Named path parameters may be set via the operation API",
-          actual,
-          is(URI.create("https://example.com/pets/1234")));
-    }
-  }
-
-  /** Lily generates fluent-style setters for all query parameters. */
-  @Nested
-  class HasFluentQueryParameters {
-    private static String packageName;
-
-    @BeforeAll
-    static void beforeAll() throws Exception {
-      packageName =
-          compileOas(
-              """
-              openapi: 3.0.2
-              paths:
-                /pets:
-                  get:
-                    operationId: listPets
-                    parameters:
-                      - name: limit
-                        in: query
-                        schema:
-                          type: integer
-                          format: int32
-                        # Path parameters are simple by default
-                        # style: simple
-                      - name: include
-                        in: query
-                        schema:
-                          type: array
-                          items:
-                            type: string
-                        # Query parameters are form-explode by default
-                        # style: form
-                        # explode: true
-              """);
-    }
-
-    @Test
-    void hasQueryParameterSetters() {
-      var actual =
-          evaluate(
-              """
-              return %s.Api.newBuilder()
-                .uri("https://example.com/")
-                .build()
-                .everyOperation()
-                .listPets()
-                .setLimit(5)
-                .setInclude(java.util.List.of("name", "age"))
-                .uriTemplate()
-                .toURI();
-              """
-                  .formatted(packageName),
-              URI.class);
-      assertThat(
-          "Query parameters may be set via the operation API",
-          actual,
-          is(URI.create("https://example.com/pets?include=name&include=age&limit=5")));
-    }
-  }
 }
